+++ CSE 514 Networking/P2P Application/TcpTorrent/TcpTorrent/StateObject.cs
@@ -11,18 +11,18 @@ namespace TcpTorrent
         public string Address = string.Empty;
         public int Port = 0;
         public List<string> UploadableFilePath = new List<string>();
-        public List<int> UploadableFileSize = new List<int>();
+        public List<long> UploadableFileSize = new List<long>();
         public String TempFolderPath = string.Empty;
 
         public List<string> FilePathsToReg = new List<string>();
-        public List<int> FilePathsToRegLength = new List<int>();
+        public List<long> FilePathsToRegLength = new List<long>();
         public List<string> DownloadableFileName = new List<string>();
-        public List<int> DownloadableFileSize = new List<int>();
+        public List<long> DownloadableFileSize = new List<long>();
         public int MaxChunkSize = 8000;
         public bool ClientType = true;
         public Dictionary<string, dynamic> FileDict = new Dictionary<string, dynamic>();
         public string FileNameToDownload = string.Empty;
-        public int FileNameToDownloadLength = 0;
+        public long FileNameToDownloadLength = 0;
 
     }
 }
